# api/openapi.yaml
openapi: 3.0.3
info:
  title: Syncphony API
  version: 0.1.0
  description: |
    Shared, fair music queue across streaming services.
    This file is the source of truth for the HTTP API. Go server types and the
    web client's TypeScript types are generated from it (`make gen`).
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /api
security:
  - session: []
paths:
...
  /invites/{code}:
    parameters:
      - $ref: "#/components/parameters/InviteCode"
    get:
      operationId: getInvite
      summary: Check an invite before signing up
      description: Returns 404 if the invite doesn't exist, was used, or expired.
      tags: [auth]
      security: []
      responses:
        "200":
          description: The invite can be used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteInfo"
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteInvite
      summary: Revoke an invite (admin)
      tags: [admin]
      responses:
        "204":
          description: Revoked
        default:
          $ref: "#/components/responses/Error"
...
  /auth/signup/passkey/begin:
    post:
      operationId: beginPasskeySignup
      summary: Start creating an account with an invite and a passkey
      description: |
        Returns WebAuthn creation options for `navigator.credentials.create`.
        The account is created when the ceremony finishes.
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasskeySignupRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ceremony"
        default:
          $ref: "#/components/responses/Error"
  /auth/signup/passkey/finish:
    post:
      operationId: finishPasskeySignup
      summary: Finish creating an account with a passkey
      description: Signs the new user in.
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinishCeremony"
      responses:
        "201":
          $ref: "#/components/responses/SignedIn"
        default:
          $ref: "#/components/responses/Error"
...
  /auth/passkey/begin:
    post:
      operationId: beginPasskeyLogin
      summary: Start signing in with a passkey
      description: Returns WebAuthn request options for `navigator.credentials.get`. No username needed.
      tags: [auth]
      security: []
      responses:
        "200":
          $ref: "#/components/responses/Ceremony"
        default:
          $ref: "#/components/responses/Error"
  /auth/passkey/finish:
    post:
      operationId: finishPasskeyLogin
      summary: Finish signing in with a passkey
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinishCeremony"
      responses:
        "200":
          $ref: "#/components/responses/SignedIn"
        default:
          $ref: "#/components/responses/Error"
...
  /links/oauth/callback:
    get:
      operationId: completeOAuthLink
      summary: OAuth2 redirect target (browser navigation)
      description: |
        Finishes linking and redirects to `/settings/services?linked=<link id>`,
        or `/settings/services?link_error=<error code>` on failure. Needs the
        session cookie, but always redirects rather than returning an error.
      tags: [links]
      security: []
      parameters:
        - name: state
          in: query
          schema:
            type: string
        - name: code
          in: query
          schema:
            type: string
        - name: error
          in: query
          description: Set by the service when the user declined.
          schema:
            type: string
      responses:
        "303":
          description: Back to the web app
          headers:
            Location:
              schema:
                type: string

  # --- Browsing your services ----------------------------------------------
...
  /display/pairing:
    post:
      operationId: beginDisplayPairing
      summary: Start pairing this device as a display
      description: |
        Returns a short code for the display to show, and sets a pairing
        cookie. Poll `GET /display/pairing` until someone in a room types the
        code in. Codes last 10 minutes.
      tags: [displays]
      security: []
      responses:
        "201":
          description: A code to show
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisplayPairing"
        default:
          $ref: "#/components/responses/Error"
    get:
      operationId: pollDisplayPairing
      summary: Check whether this display has been paired
      description: |
        Reads the pairing cookie. Once paired, sets the display cookie and
        returns the room. `pairing_expired` (410) when the code timed out:
        start again.
      tags: [displays]
      security: []
      responses:
        "200":
          description: The pairing's status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisplayPairingStatus"
        default:
          $ref: "#/components/responses/Error"
  /display:
    get:
      operationId: getDisplay
      summary: This display, and its room
      description: Reads the display cookie. `unauthenticated` (401) if this device isn't a paired display.
      tags: [displays]
      security: []
      responses:
        "200":
          description: The display
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisplayMe"
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: leaveDisplay
      summary: Unpair this display, from the display
      tags: [displays]
      security: []
      responses:
        "204":
          description: Unpaired; the display cookie is cleared
        default:
          $ref: "#/components/responses/Error"
...
  /guest/{token}:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getGuestInvite
      summary: Check a guest pass before joining
      description: |
        `guest_pass_invalid` (404) if it's wrong, expired or revoked, or
        the room no longer allows guests. Wrong passes are rate limited by IP.
      tags: [guests]
      security: []
      responses:
        "200":
          description: The room the pass joins
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestInvite"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: joinAsGuest
      summary: Join a room as a guest
      description: |
        Makes a guest account and signs it in until the pass expires. A
        guest has no password, passkey, services or settings. They can
        search the server's shared services, add songs to their own lane
        (up to the room's limit; `guest_limit`, 409, past it), react, and
        vote and heart if the room allows it, in this room only. When they
        expire or are removed, their waiting songs leave the queue; what
        they played stays in the room's history under their name.
      tags: [guests]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinAsGuestRequest"
      responses:
        "201":
          $ref: "#/components/responses/SignedIn"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: syncphony_session
      description: |
        Set by the sign-in endpoints. httpOnly, SameSite=Lax, and Secure when
        the server is served over HTTPS. Mutating requests must come from the
        app's own origin (checked with Sec-Fetch-Site / Origin).
    display:
      type: apiKey
      in: cookie
      name: syncphony_display
      description: |
        A paired display's token (see `/display/pairing`). It can read its
        own room and nothing else: the operations that list it here, and
        the room socket.
  parameters:
    InviteCode:
      name: code
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Error:
      description: |
        An error. Common statuses: 400 invalid input, 401 not signed in or
        wrong credentials, 403 not allowed, 404 not found, 409 conflict,
        429 too many attempts (see Retry-After).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    SignedIn:
      description: Signed in; the session cookie is set
      headers:
        Set-Cookie:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Me"
    Ceremony:
      description: A WebAuthn ceremony to complete in the browser
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Ceremony"
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable, machine-readable error code.
          example: invalid_credentials
        message:
          type: string
          description: Human-readable explanation.
    Role:
      type: string
      enum: [admin, member]
    User:
      type: object
      required: [id, username, displayName, color, role, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatar:
          type: string
          description: Image URL. Absent means show initials on `color`.
        color:
          type: string
          description: Lane color, `#rrggbb`.
          example: "#7c3aed"
        role:
          $ref: "#/components/schemas/Role"
        guest:
          $ref: "#/components/schemas/UserGuest"
        createdAt:
          type: string
          format: date-time
    Me:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          required: [hasPassword, passkeyCount]
          properties:
            hasPassword:
              type: boolean
            passkeyCount:
              type: integer
    UserGuest:
      type: object
      description: Set on guests, who joined one room with a QR code instead of an invite.
      required: [roomId, expiresAt, ended]
      properties:
        roomId:
          type: string
        expiresAt:
          type: string
          format: date-time
        ended:
          type: boolean
          description: |
            The guest expired or was removed. They're kept only so the room's
            history and recaps can show their name.
    RoomGuests:
      type: object
      description: |
        Whether people without an account may join the room by scanning a
        guest pass, and what they may do.
      required: [allowed, maxSongs, canVote]
      properties:
        allowed:
          type: boolean
          default: false
        maxSongs:
          type: integer
          minimum: 0
          maximum: 100
          default: 10
          description: Songs each guest may add over their visit. 0 means no limit.
        canVote:
          type: boolean
          default: true
          description: Guests may vote to skip and heart songs.
...
    GuestInvite:
      type: object
      required: [roomId, roomName, expiresAt]
      properties:
        roomId:
          type: string
        roomName:
          type: string
        expiresAt:
          type: string
          format: date-time
    JoinAsGuestRequest:
      type: object
      required: [displayName]
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 32
...
    Username:
      type: string
      description: "2 to 32 characters: lowercase letters, digits, `.`, `_` and `-`."
      pattern: "^[a-z0-9._-]{2,32}$"
...
    PasskeySignupRequest:
      type: object
      required: [invite, username, displayName]
      properties:
        invite:
          type: string
        username:
          $ref: "#/components/schemas/Username"
        displayName:
          type: string
          minLength: 1
          maxLength: 64
...
    Ceremony:
      type: object
      required: [ceremonyId, options]
      properties:
        ceremonyId:
          type: string
          description: Pass back to the matching finish endpoint within 5 minutes.
        options:
          type: object
          additionalProperties: true
          description: |
            WebAuthn options as JSON (`publicKey` inside). Decode with
            `PublicKeyCredential.parseCreationOptionsFromJSON` or
            `parseRequestOptionsFromJSON`.
    FinishCeremony:
      type: object
      required: [ceremonyId, credential]
      properties:
        ceremonyId:
          type: string
        credential:
          type: object
          additionalProperties: true
          description: The browser's `PublicKeyCredential.toJSON()`.
        name:
          type: string
          maxLength: 64
          description: Label for a new passkey, e.g. "Pixel 9". Optional.
...
    InviteInfo:
      type: object
      required: [role, expiresAt]
      properties:
        role:
          $ref: "#/components/schemas/Role"
        expiresAt:
          type: string
          format: date-time
...
    Display:
      type: object
      required: [id, roomId, name, createdAt, lastSeenAt]
      properties:
        id:
          type: string
        roomId:
          type: string
        name:
          type: string
        pairedBy:
          type: string
          description: The user who paired it, if they're still around.
        createdAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
...
    DisplayPairing:
      type: object
      required: [code, expiresAt]
      properties:
        code:
          type: string
        expiresAt:
          type: string
          format: date-time
    DisplayPairingStatus:
      type: object
      required: [status, code, expiresAt]
      properties:
        status:
          type: string
          enum: [waiting, paired]
        code:
          type: string
        expiresAt:
          type: string
          format: date-time
        roomId:
          type: string
          description: Set once paired.
    DisplayMe:
      type: object
      required: [display, room]
      properties:
        display:
          $ref: "#/components/schemas/Display"
        room:
          $ref: "#/components/schemas/Room"
...
    FairnessMode:
      type: string
      enum: [round_robin, fifo]
    PermissionLevel:
      type: string
      enum: [everyone, owner]
      description: Who may do something. The room's owner always may.
    SkipPermission:
      type: string
      enum: [everyone, vote, owner]
      description: |
        Who may skip. `vote`: the owner may, and everyone else votes (see
        SkipVotes). Whoever queued a song may always skip it.
    RoomPermissions:
      type: object
      required: [playPause, seek, skip, speaker]
      properties:
        playPause:
          $ref: "#/components/schemas/PermissionLevel"
        seek:
          $ref: "#/components/schemas/PermissionLevel"
        skip:
          $ref: "#/components/schemas/SkipPermission"
        speaker:
          $ref: "#/components/schemas/PermissionLevel"
...
    SkipVotePercent:
      type: integer
      minimum: 0
      maximum: 99
      default: 50
      description: |
        A skip vote passes once more than this percent of the room has
        voted (not counting whoever queued the song): 50 is a majority.
    RoomFairness:
      type: object
      description: |
        Tunes the room's fairness mode. While someone else has songs
        waiting, nobody gets more than `maxInARow` songs in a row, and
        `cooldown` other songs play between one person's songs; when
        nobody else is waiting, the music keeps going anyway. In round
        robin, `weights` give some people more songs per turn.
      required: [maxInARow, cooldown, weights, repeatWindowMinutes]
      properties:
        maxInARow:
          type: integer
          minimum: 0
          maximum: 10
          description: 0 is no cap.
        cooldown:
          type: integer
          minimum: 0
          maximum: 10
          description: Songs by others between one person's songs. 0 is none.
        weights:
          type: object
          description: Songs per turn (2 to 4) by user ID, in round robin. Everyone else gets 1.
          additionalProperties:
            type: integer
            minimum: 1
            maximum: 4
        repeatWindowMinutes:
          type: integer
          minimum: 0
          maximum: 1440
          description: |
            Refuse songs already waiting or playing, or started within this
            many minutes. 0 is off.
    RoomMatching:
      type: object
      description: How the room uses the same song on other services.
      required: [fallback, borrow]
      properties:
        fallback:
          type: boolean
          default: true
          description: |
            When a song's service can't play it, play the same recording
            from another service in the room: a service of someone who's
            here, or a shared one. Matched by ISRC, or carefully by title,
            artist and length.
        borrow:
          type: boolean
          default: false
          description: |
            Anyone may queue a song the room played again, even if it's only
            on someone else's service.
    RoomAutopilot:
      type: object
      description: |
        When the queue runs dry, autopilot adds songs like the ones the
        room has been playing, taking seeds from each member's songs in
        turn. It only plays while the room has a speaker.
      required: [on, adventure]
      properties:
        on:
          type: boolean
          default: false
        adventure:
          type: string
          enum: [similar, discovery]
          default: similar
          description: |
            `similar` stays close to the room's songs, their artists
            included. `discovery` plays other artists, further afield.
    Room:
      type: object
      required: [id, name, ownerId, fairnessMode, fairness, matching, autopilot, guests, permissions, skipVotePercent, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        ownerId:
          type: string
        fairnessMode:
          $ref: "#/components/schemas/FairnessMode"
        fairness:
          $ref: "#/components/schemas/RoomFairness"
        matching:
          $ref: "#/components/schemas/RoomMatching"
        autopilot:
          $ref: "#/components/schemas/RoomAutopilot"
        guests:
          $ref: "#/components/schemas/RoomGuests"
        permissions:
          $ref: "#/components/schemas/RoomPermissions"
        skipVotePercent:
          $ref: "#/components/schemas/SkipVotePercent"
        createdAt:
          type: string
          format: date-time
...